09 02 2026 [ update produto]

+++ src/controllers/empresa/empresaController.ts
@@ -95,7 +95,7 @@ export class CreateEmpresa {
           codigo int(11) unsigned NOT NULL AUTO_INCREMENT,
         id  int(10) unsigned NOT NULL DEFAULT 0,
         estoque  double DEFAULT 0,
-        preco  double DEFAULT 0,
+        preco  decimal(10,2) DEFAULT 0.00,
         unidade_medida  varchar(255) DEFAULT 'UND',
         grupo  int(11) DEFAULT 0,
         origem  char(1) NOT NULL DEFAULT '0',
